[host] Refactor/reduce `main` code
