+++ src/emulator/host/include/host/app.h
@@ -17,6 +17,9 @@
 
 #pragma once
 
+#include <glbinding/gl/types.h>
+
+#include <memory>
 #include <string>
 #include <vector>
 
@@ -28,8 +31,28 @@ static constexpr auto WINDOW_BORDER_HEIGHT = 34;
 typedef std::vector<uint8_t> Buffer;
 
 struct HostState;
+struct SDL_Window;
+struct SDL_Surface;
 template <class T>
 class Ptr;
 
-bool load_app(Ptr<const void> &entry_point, HostState &host, const std::wstring &path, bool is_vpk);
+typedef std::vector<uint8_t> Buffer;
+typedef std::unique_ptr<const void, void(*)(const void *)> SDLPtr;
+typedef std::unique_ptr<SDL_Surface, void(*)(SDL_Surface *)> SurfacePtr;
+
+enum ExitCode {
+    Success = 0,
+    IncorrectArgs,
+    SDLInitFailed,
+    HostInitFailed,
+    ModuleLoadFailed,
+    InitThreadFailed,
+    RunThreadFailed
+};
+
+void error_dialog(const std::string &message, SDL_Window *window = nullptr);
+ExitCode load_app(Ptr<const void> &entry_point, HostState &host, const std::wstring &path, bool is_vpk);
+ExitCode run_app(HostState &host, Ptr<const void> &entry_point);
 bool read_file_from_disk(Buffer &buf, const char *file, HostState &host);
+void set_window_title(HostState &host);
+void no_fb_fallback(HostState &host, gl::GLuint *fb_texture_id);
