+++ src/emulator/host/src/app.cpp
@@ -22,8 +22,11 @@
 #include <host/functions.h>
 #include <host/sfo.h>
 #include <host/state.h>
+#include <host/version.h>
+#include <kernel/thread/thread_functions.h>
 #include <io/state.h>
 #include <util/fs.h>
+#include <util/find.h>
 #include <util/log.h>
 #include <util/string_convert.h>
 
@@ -33,6 +36,7 @@
 #include <cassert>
 #include <cstring>
 #include <fstream>
+#include <sstream>
 #include <iostream>
 #include <istream>
 #include <iterator>
@@ -42,6 +46,14 @@
 #include <gui/imgui_impl_sdl_gl3.h>
 // clang-format on
 #include <gui/functions.h>
+
+using namespace gl;
+
+void error_dialog(const std::string &message, SDL_Window *window) {
+    if (SDL_ShowSimpleMessageBox(SDL_MESSAGEBOX_ERROR, "Error", message.c_str(), window) < 0) {
+        LOG_ERROR("SDL Error: {}", message);
+    }
+}
 
 static void delete_zip(mz_zip_archive *zip) {
     mz_zip_reader_end(zip);
@@ -190,7 +202,7 @@ bool install_vpk(Ptr<const void> &entry_point, HostState &host, const std::wstri
     return true;
 }
 
-bool load_app(Ptr<const void> &entry_point, HostState &host, const std::wstring &path, bool is_vpk) {
+bool load_app_impl(Ptr<const void> &entry_point, HostState &host, const std::wstring &path, bool is_vpk) {
     if (is_vpk) {
         if (!install_vpk(entry_point, host, path)) {
             return false;
@@ -234,4 +246,72 @@ bool load_app(Ptr<const void> &entry_point, HostState &host, const std::wstring 
     }
 
     return true;
+}
+
+ExitCode load_app(Ptr<const void> &entry_point, HostState &host, const std::wstring &path, bool is_vpk) {
+    if (!load_app_impl(entry_point, host, path, is_vpk)) {
+        std::string message = "Failed to load \"";
+        message += wide_to_utf(path);
+        message += "\"";
+        message += "\nSee console output for details.";
+        error_dialog(message.c_str(), host.window.get());
+        return ModuleLoadFailed;
+    }
+
+    return Success;
+}
+
+
+ExitCode run_app(HostState &host, Ptr<const void> &entry_point) {
+    const CallImport call_import = [&host](uint32_t nid, SceUID main_thread_id) {
+        ::call_import(host, nid, main_thread_id);
+    };
+
+    const SceUID main_thread_id = create_thread(entry_point, host.kernel, host.mem, host.io.title_id.c_str(), SCE_KERNEL_DEFAULT_PRIORITY_USER, SCE_KERNEL_STACK_SIZE_USER_MAIN, call_import, false);
+    if (main_thread_id < 0) {
+        error_dialog("Failed to init main thread.", host.window.get());
+        return InitThreadFailed;
+    }
+
+    const ThreadStatePtr main_thread = find(main_thread_id, host.kernel.threads);
+    Ptr<void> argp = Ptr<void>();
+    if (!strncmp(host.kernel.loaded_modules.begin()->second->module_name, "SceLibc", 7)) {
+        const SceUID libc_thread_id = create_thread(host.kernel.loaded_modules.begin()->second->module_start, host.kernel, host.mem, "libc", SCE_KERNEL_DEFAULT_PRIORITY_USER, SCE_KERNEL_STACK_SIZE_USER_DEFAULT, call_import, false);
+        const ThreadStatePtr libc_thread = find(libc_thread_id, host.kernel.threads);
+        run_on_current(*libc_thread, host.kernel.loaded_modules.begin()->second->module_start, 0, argp);
+        libc_thread->to_do = ThreadToDo::exit;
+        libc_thread->something_to_do.notify_all(); // TODO Should this be notify_one()?
+        host.kernel.running_threads.erase(libc_thread_id);
+        host.kernel.threads.erase(libc_thread_id);
+    }
+
+    if (start_thread(host.kernel, main_thread_id, 0, Ptr<void>()) < 0) {
+        error_dialog("Failed to run main thread.", host.window.get());
+        return RunThreadFailed;
+    }
+
+    return Success;
+}
+
+void no_fb_fallback(HostState &host, GLuint *fb_texture_id)
+{
+    glGenTextures(1, fb_texture_id);
+    glClearColor(1.0, 0.0, 0.5, 1.0);
+    glClearDepth(1.0f);
+    glViewport(0, 0, host.display.image_size.width, host.display.image_size.height);
+}
+
+void set_window_title(HostState& host)
+{
+    const uint32_t sdl_ticks_now = SDL_GetTicks();
+    const uint32_t ms = sdl_ticks_now - host.sdl_ticks;
+    if (ms >= 1000 && host.frame_count > 0) {
+        const uint32_t fps = (host.frame_count * 1000) / ms;
+        const uint32_t ms_per_frame = ms / host.frame_count;
+        std::ostringstream title;
+        title << window_title << " | " << host.game_title << " (" << host.io.title_id << ") | " << ms_per_frame << " ms/frame (" << fps << " frames/sec)";
+        SDL_SetWindowTitle(host.window.get(), title.str().c_str());
+        host.sdl_ticks = sdl_ticks_now;
+        host.frame_count = 0;
+    }
 }
