+++ src/emulator/main.cpp
@@ -20,10 +20,7 @@
 #include <host/state.h>
 #include <host/version.h>
 #include <kernel/thread/thread_functions.h>
-#include <kernel/thread/thread_state.h>
-#include <util/find.h>
 #include <util/log.h>
-#include <util/resource.h>
 #include <util/string_convert.h>
 
 #include <SDL.h>
@@ -31,40 +28,9 @@
 
 #include <algorithm> // find_if_not
 #include <cassert>
-#include <iostream>
-#include <sstream>
 
 #include <gui/functions.h>
 #include <gui/imgui_impl.h>
-
-typedef std::unique_ptr<const void, void (*)(const void *)> SDLPtr;
-typedef std::unique_ptr<SDL_Surface, void (*)(SDL_Surface *)> SurfacePtr;
-
-enum ExitCode {
-    Success = 0,
-    IncorrectArgs,
-    SDLInitFailed,
-    HostInitFailed,
-    ModuleLoadFailed,
-    InitThreadFailed,
-    RunThreadFailed
-};
-
-static bool is_macos_process_arg(const char *arg) {
-    return strncmp(arg, "-psn_", 5) == 0;
-}
-
-static void error(const std::string &message, SDL_Window *window = nullptr) {
-    if (SDL_ShowSimpleMessageBox(SDL_MESSAGEBOX_ERROR, "Error", message.c_str(), window) < 0) {
-        LOG_ERROR("SDL Error: {}", message);
-    }
-}
-
-static void term_sdl(const void *succeeded) {
-    assert(succeeded != nullptr);
-
-    SDL_Quit();
-}
 
 int main(int argc, char *argv[]) {
     init_logging();
@@ -73,15 +39,22 @@ int main(int argc, char *argv[]) {
 
     ProgramArgsWide argv_wide = process_args(argc, argv);
 
-    const SDLPtr sdl(reinterpret_cast<const void *>(SDL_Init(SDL_INIT_AUDIO | SDL_INIT_GAMECONTROLLER | SDL_INIT_VIDEO) >= 0), term_sdl);
+    const SDLPtr sdl(reinterpret_cast<const void *>(SDL_Init(SDL_INIT_AUDIO | SDL_INIT_GAMECONTROLLER | SDL_INIT_VIDEO) >= 0), [](const void *succeeded) {
+        assert(succeeded != nullptr);
+        SDL_Quit();
+    });
     if (!sdl) {
-        error("SDL initialisation failed.");
+        error_dialog("SDL initialisation failed.");
         return SDLInitFailed;
     }
 
     SDL_GL_SetAttribute(SDL_GL_DOUBLEBUFFER, 1);
 
-    const char *const *const path_arg = std::find_if_not(&argv[1], &argv[argc], is_macos_process_arg);
+    // Filter out an argument that macOS Finder appends 
+    const char *const *const path_arg = std::find_if_not(&argv[1], &argv[argc], [](const char *arg) {
+        return strncmp(arg, "-psn_", 5) == 0;
+    });
+
     std::wstring path;
     if (path_arg != &argv[argc]) {
         path = utf_to_wide(*path_arg);
@@ -124,50 +97,17 @@ int main(int argc, char *argv[]) {
     }
 
     Ptr<const void> entry_point;
-    if (!load_app(entry_point, host, path, is_vpk)) {
-        std::string message = "Failed to load \"";
-        message += wide_to_utf(path);
-        message += "\"";
-        message += "\nSee console output for details.";
-        error(message.c_str(), host.window.get());
-        return ModuleLoadFailed;
-    }
+    if (auto err = load_app(entry_point, host, path, is_vpk) != Success)
+        return err;
 
-    const CallImport call_import = [&host](uint32_t nid, SceUID main_thread_id) {
-        ::call_import(host, nid, main_thread_id);
-    };
+    if (auto err = run_app(host, entry_point) != Success)
+        return err;
 
-    const SceUID main_thread_id = create_thread(entry_point, host.kernel, host.mem, host.io.title_id.c_str(), SCE_KERNEL_DEFAULT_PRIORITY_USER, SCE_KERNEL_STACK_SIZE_USER_MAIN, call_import, false);
-    if (main_thread_id < 0) {
-        error("Failed to init main thread.", host.window.get());
-        return InitThreadFailed;
-    }
-
-    const ThreadStatePtr main_thread = find(main_thread_id, host.kernel.threads);
-    Ptr<void> argp = Ptr<void>();
-    if (!strncmp(host.kernel.loaded_modules.begin()->second->module_name, "SceLibc", 7)) {
-        const SceUID libc_thread_id = create_thread(host.kernel.loaded_modules.begin()->second->module_start, host.kernel, host.mem, "libc", SCE_KERNEL_DEFAULT_PRIORITY_USER, SCE_KERNEL_STACK_SIZE_USER_DEFAULT, call_import, false);
-        const ThreadStatePtr libc_thread = find(libc_thread_id, host.kernel.threads);
-        run_on_current(*libc_thread, host.kernel.loaded_modules.begin()->second->module_start, 0, argp);
-        libc_thread->to_do = ThreadToDo::exit;
-        libc_thread->something_to_do.notify_all(); // TODO Should this be notify_one()?
-        host.kernel.running_threads.erase(libc_thread_id);
-        host.kernel.threads.erase(libc_thread_id);
-    }
-
-    if (start_thread(host.kernel, main_thread_id, 0, Ptr<void>()) < 0) {
-        error("Failed to run main thread.", host.window.get());
-        return RunThreadFailed;
-    }
-
-    GLuint TextureID = 0;
-    host.t1 = SDL_GetTicks();
+    GLuint fb_texture_id = 0;
+    host.sdl_ticks = SDL_GetTicks();
     while (handle_events(host)) {
-        if (!TextureID) {
-            glGenTextures(1, &TextureID);
-            glClearColor(1.0, 0.0, 0.5, 1.0);
-            glClearDepth(1.0f);
-            glViewport(0, 0, DEFAULT_RES_WIDTH, DEFAULT_RES_HEIGHT);
+        if (!fb_texture_id) {
+            no_fb_fallback(host, &fb_texture_id);
         }
 
         imgui::draw_begin(host.window);
@@ -181,17 +121,7 @@ int main(int argc, char *argv[]) {
 
         host.display.condvar.notify_all();
 
-        const uint32_t t2 = SDL_GetTicks();
-        const uint32_t ms = t2 - host.t1;
-        if (ms >= 1000 && host.frame_count > 0) {
-            const uint32_t fps = (host.frame_count * 1000) / ms;
-            const uint32_t ms_per_frame = ms / host.frame_count;
-            std::ostringstream title;
-            title << window_title << " | " << host.game_title << " (" << host.io.title_id << ") | " << ms_per_frame << " ms/frame (" << fps << " frames/sec)";
-            SDL_SetWindowTitle(host.window.get(), title.str().c_str());
-            host.t1 = t2;
-            host.frame_count = 0;
-        }
+        set_window_title(host);
     }
     return Success;
 }
